requirements.txt
# FastAPI and server
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
python-multipart>=0.0.12,<0.0.13

# Database
sqlalchemy>=2.0.36,<2.1.0
psycopg2-binary>=2.9.10,<2.10.0
alembic>=1.14.0,<1.15.0

# Configuration and validation
pydantic-settings>=2.6.0,<2.7.0
python-dotenv>=1.0.1,<2.0.0

# HTTP client for Ollama (local LLM)
httpx>=0.27.0,<0.28.0

# Agent orchestration (grounded exploitability)
langgraph>=0.2.0,<0.3.0
langchain-core>=0.3.0,<0.4.0

# Auth
PyJWT>=2.8.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
